requirements.txt
Django
djangorestframework
coreapi

# test
pytest
model_mommy

# dev
isort
yapf
